Fixed a bug that led newly introduced hex decoding function not to work in eci-load*.c (ushort instead of uint was used for months). Changes to the usage() functions, removed a dirty debugging exit(0) in pppoeci.c (gosh!).

+++ usermode/util.h
@@ -4,10 +4,10 @@
 
 struct config_t
 {
-	unsigned short vid1;
-	unsigned short pid1;
-	unsigned short vid2;
-	unsigned short pid2;
+	unsigned int vid1;
+	unsigned int pid1;
+	unsigned int vid2;
+	unsigned int pid2;
 	unsigned int vpi;
 	unsigned int vci;
 	char* mode;
